#9 Refactor APRS Collection logic
This is quie a deep change. Collecting logic has been moved to individual objects/classes. This reduces amount of code (better maitainability) and prepares the path for supporting messages.
Supporting newer slow data types is just a matter of adding new ClowDataCollector.
This also brings the benefit to support interleaved slow data as this is the case with RS-MS1A messages (messages are interleaved with GPS Data)
The source callsign for NMEA sentences is also set with correct SSID (-5) as per http://www.aprs.org/aprs11/SSIDs.txt
The NMEA sentences are no longer converted but sent to APRS-S almost as is, just packed into an APRS frame.